renamed to LayoutChangeForTFNCHWConvolutions.py
